pyproject.toml
[project]
name = "gencon"
version = "0.1.0"
description = "Genie-Force - AI/BI Genie Room Manager for Databricks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.29.0",
    "databricks-langchain>=0.4.0",
    "databricks-sdk>=0.30.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langgraph>=0.2.0",
    "langgraph-supervisor>=0.0.30",
    "openai>=1.52.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
]
